# rand: Use stdlib crypto/rand.Read on Linux 6.11 + Go 1.24