@jrick 2024-12-17
Unfortunately it looks like reading 32 bytes took a small performance hit.